### 2023-02-17 #195 @whedon
[PRE REVIEW]: R for Data Analysis: An open-source resource for teaching and learning analytics with R
**Submitting author:** @TrevorFrench (<a href="http://orcid.org/0000-0002-6246-2249">Trevor French</a>)
**Repository:** <a href="https://github.com/TrevorFrench/R-for-Data-Analysis" target ="_blank">https://github.com/TrevorFrench/R-for-Data-Analysis</a>
**Version:** 1.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
