### 2023-03-02 #195 @stats-tgeorge
Hello @TrevorFrench. Looks like you have included the necessary pieces for submission and your paper is in the JOSE scope. I will start to identify reviewers!
