### 2023-03-09 #195 @nhejazi
@stats-tgeorge, this seems like a very thoughtfully written book and I would normally be eager to provide some input; however, I am currently facing a host of approaching deadlines and will not be able to provide a timely review -- my apologies for this.
