### 2023-03-10 #195 @neuroimaginador
I'm so sorry, but at this moment I will not be able to review this
submission (although it seems quite interesting) due to other duties.

El mié, 8 mar 2023 a las 17:21, Tyler George ***@***.***>)
escribió:

> Hello, @neuroimaginador <https://github.com/neuroimaginador>, you signed
> up to review for JOSE. Would you be willing to review for a submission
> titled "R for Data Analysis: An open-source resource for teaching and
> learning analytics with R”? Let me know if you're available and willing.
> Thank you!
>
