### 2023-03-10 #195 @stats-tgeorge
Hello, @Stomsing1, you signed up to review for JOSE. Would you be willing to review for a submission titled "R for Data Analysis: An open-source resource for teaching and learning analytics with R”? Let me know if you're available and willing. Thank you!
