### 2023-03-13 #195 @stats-tgeorge
@tomsing1  Wonderful thank you. I'm still looking for at least 1 more reviewer before the whole process gets started. Please read through the reviewer documentation (if you are not familiar) when you ave some time. Available [here](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html).
