### 2023-03-13 #195 @stats-tgeorge
Hello, @corybrunson, you signed up to review for JOSE. Would you be willing to review for a submission titled "R for Data Analysis: An open-source resource for teaching and learning analytics with R”? Let me know if you're available and willing. Thank you!
