### 2023-03-13 #195 @corybrunson
Hi @stats-tgeorge, i may be able to, though it may take longer than a typical software package review. What is the expected or preferred turnaround for a book-length resource such as this?
