### 2023-03-14 #195 @njtierney
Hi @stats-tgeorge - I would be happy to however I am on leave for the next 2 weeks, back on the 29th March - I'll be free from the 29th March until April 27th to review.
