### 2023-03-14 #195 @stats-tgeorge
@corybrunson around a month? We don't review a lot of resources similar to this. Does that sound like ti would work for you?
