### 2023-03-14 #202 @whedon
[REVIEW]: R for Data Analysis: An open-source resource for teaching and learning analytics with R
**Submitting author:** @TrevorFrench (<a href="http://orcid.org/0000-0002-6246-2249">Trevor French</a>)
**Repository:** <a href="https://github.com/TrevorFrench/R-for-Data-Analysis" target ="_blank">https://github.com/TrevorFrench/R-for-Data-Analysis</a>
**Version:** v1.1.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".

* [x]  **Repository:** Is the source for this learning module available at the [repository url](https://github.com/TrevorFrench/R-for-Data-Analysis)?
  * [raised issue](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/10)

  * This is an online book, and various R packages are referenced and introduced throughout.
   * The online book appears to be mainly intended for self-study, not for use in a classroom setting. It includes numerous code examples and screenshots, but does not discuss how a teacher might adapt it e.g. for a class.
  * [raised issue](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/16)

  * missing affiliations [raised issue](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/11)
  * [raised issue](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/18)
online books.
  * [raised issue](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/19)
