### 2023-04-02 TrevorFrench/R-for-Data-Analysis#10 @tomsing1
JOSE review - plain text LICENSE file
Hi!

Great course / e-book!
I am one of the reviewers of your submission to the Journal of Open Source Education. As I am going through their checklist, I will create issues here that you might want to take a look at.

For example, JOSE asks me to verify:

> **License:** Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)

You are already [linking to the CC license](https://trevorfrench.github.io/R-for-Data-Analysis/#license). Would you mind adding the LICENSE text to the repository as well?
