### 2023-04-02 TrevorFrench/R-for-Data-Analysis#11 @tomsing1
JOSE review - paper.md is missing the affiliations field
It seems that the JOSE bot cannot create the PDF version of the `paper.md` file because the `affiliations` field is missing the YAML header (see [here](https://github.com/openjournals/jose-reviews/issues/202#issuecomment-1468493764)). Would you mind adding it?

[Here is an example](https://raw.githubusercontent.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/main/paper/paper.md) of another manuscript (currently under review) that includes several affiliations.
