### 2023-04-02 #195 @labarba
Interesting... I notice that our bot added @tomsing1 twice to the list of reviewers at the top of this thread.

I didn't know that could happen! As the reviewer list carries over to the published paper, I'll try removing the re-inviting...

@stats-tgeorge We also have a command `@whedon re-invite … as reviewer` for when the invitation expires.
