### 2023-04-02 c4267c7
Fixed typos, added a sentence about NA, added reference to the corresponding chapter in Hadley's Advanced R book.
