### 2023-04-03 TrevorFrench/R-for-Data-Analysis#10 @TrevorFrench
Hi- thanks for agreeing to review!

I added the LICENSE file to this repository. If that looks alright to you, I'll close this issue.
