### 2023-04-03 TrevorFrench/R-for-Data-Analysis#11 @TrevorFrench
I updated the `paper.md` file to include affiliations (thanks for sharing an example); however, I am not affiliated with any organization. I listed my affiliation as "Independent". Do you think that is sufficient?
