### 2023-04-06 TrevorFrench/R-for-Data-Analysis#11 @tomsing1
@TrevorFrench I just saw that [the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) contains an example for independent researchers like yourself. It suggests `Independent Researcher`, `Country`.  Perhaps that works for you, too?
