### 2023-04-06 TrevorFrench/R-for-Data-Analysis#11 @TrevorFrench
@tomsing1 Oh, perfect! I just updated it to be "Independent Researcher, USA" in PR #13
