### 2023-04-08 #202 @corybrunson
@stats-tgeorge could you please clarify whether the "Documentation" section of the checklist refers to the repository or the website / product? My understanding is the former, but i want to be sure before marking items off or raising issues on the repo. Thank you!
