### 2023-04-08 TrevorFrench/R-for-Data-Analysis#14 @corybrunson
JOSE review – Documentation checklist
Hi @TrevorFrench and thanks for the thorough resource—i'm enjoying reading it.

Following [discussion in the review thread](https://github.com/openjournals/jose-reviews/issues/202#issuecomment-1500922289), the repository—preferably the README—should include the following items from the Documentation checklist, most of which i think are adequately addressed; the exception is a statement of need. Would you be able to add some text (i think a sentence or two would do) to the opening paragraph that motivates this work amidst the other resources available to your target readership?

_Part of [this JOSE review](https://github.com/openjournals/jose-reviews/issues/202)._
