### 2023-04-08 f5ebdfd
Fixed typos, suggested the use of the predict function for the linear model.
