### 2023-04-08 a8dc0af
Various suggestions and fixed typos
