### 2023-04-08 TrevorFrench/R-for-Data-Analysis#16 @tomsing1
JOSE review: Contributor guidelines
The JOSE review guidelines ask for Contributor Guidelines. Would you mind adding some, e.g. using with [use_tidy_contributing()](https://usethis.r-lib.org/reference/tidyverse.html) as a starting point? (You will likely want to edit the template according to your own needs.)
