### 2023-04-08 TrevorFrench/R-for-Data-Analysis#17 @tomsing1
JOSE review: classroom use
The JOSE review guidelines ask
> Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

It seems this material is mainly intended for self-study? Or have you (or others) applied it to a classroom setting? If not, how do you envision that might be accomplished?
