### 2023-04-08 TrevorFrench/R-for-Data-Analysis#18 @tomsing1
JOSE review: story of the authors
The JOSE review guidelines ask:
>  Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

Maybe you want to include a chapter introducing your story, background and expertise?
