### 2023-04-08 TrevorFrench/R-for-Data-Analysis#19 @tomsing1
JOSE: additional references / comparisons with existing online materials
It would be great if you could include a few more references to / a comparison with existing educational materials in the paper.
For example, Software Carpentry has a long-standing section on [R for Reproducible Scientific Analysis](https://swcarpentry.github.io/r-novice-gapminder/). How does your book compare, e.g. where are differences in scope, what are gaps you are filling?
