### 2023-04-08 TrevorFrench/R-for-Data-Analysis#18 @TrevorFrench
@tomsing1 should this go in the content or in the paper?
