### 2023-04-08 721fadd
updating statement of need and including it in the README
