+++ paper/paper.bib
@@ -35,3 +35,121 @@
   keywords = {Big data, Systematic review, Data practices, Disruptive innovation},
   abstract = {Big data has captured the interests of scholars across many disciplines over the last half a decade. Business scholars have increasingly turned their attention to the impact of this emerging phenomenon. Despite the rise in attention, our understanding of what big data is and what it means for organizations and institutional actors remains uncertain. In this study, we conduct a systematic review on “big data” across business scholarship over the past six years (2009–2014). We analyzed 219 peer-reviewed academic papers from 152 journals from the most comprehensive business literature database. We conducted the systematic review both quantitatively and qualitatively using the data analysis software NVivo10. Our results reveal several key insights about the scholarly investigation of big data, including its top benefits and challenges. Overall, we found that big data remains a fragmented, early-stage domain of research in terms of theoretical grounding, methodological diversity and empirically oriented work. These challenges serve to improve our understanding of the state of big data in contemporary research, and to further prompt scholars and decision-makers to advance future research in the most productive manner.}
 }
+
+@misc{Zimmerman,
+  author       = {Naupaka Zimmerman and
+                  Greg Wilson and
+                  Raniere Silva and
+                  Scott Ritchie and
+                  François Michonneau and
+                  Jeffrey Oliver and
+                  Harriet Dashnow and
+                  Andrew Boughton and
+                  Andy Teucher and
+                  David Mawdsley and
+                  Andrew MacDonald and
+                  Timothy Rice and
+                  Rémi Emonet and
+                  Remi Daigle and
+                  Bill Mills and
+                  Ben Bolker and
+                  Sam Penrose and
+                  Clare Sloggett and
+                  John Blischak and
+                  Timothy Eoin Moore and
+                  David Mawdsley and
+                  Jeffrey Arnold and
+                  Dave Bridges and
+                  Erin Alison Becker and
+                  Giulio Valentino Dalla Riva and
+                  Liz Ing-Simmons and
+                  Research Bazaar and
+                  Trevor Bekolay and
+                  Julia Piaskowski and
+                  Marc Sze and
+                  Martin John Hadley and
+                  Nima Hejazi and
+                  Fernando Mayer and
+                  Katrin Leinweber and
+                  Lachlan Deer and
+                  Nicholas Lesniak and
+                  Olivia Rata Burge and
+                  Paula Andrea Martinez and
+                  Ana Costa Conrado and
+                  Andris Jankevics and
+                  Jaime Ashander and
+                  Jonah Duckles and
+                  Luke Zappia and
+                  Marie-Helene Burle and
+                  Nora Mitchell and
+                  Phil Bouchet and
+                  Rayna Michelle Harris and
+                  Sebastien Renaut and
+                  Adam H. Sparks and
+                  Daniel and
+                  Dean Attali and
+                  Drew Tyre and
+                  Elise Morrison and
+                  Gordon McDonald and
+                  Ido Bar and
+                  James Mickley and
+                  Jamie McDevitt-Irwin and
+                  Katherine Koziar and
+                  Kieran Samuk and
+                  Kunal Marwaha and
+                  Kyriakos Chatzidimitriou and
+                  Lucy Chang and
+                  Melissa Kardish and
+                  Nicholas Potter and
+                  Philipp Boersch-Supan and
+                  Scott Allen Funkhouser and
+                  Tobin Magle and
+                  waiteb5 and
+                  Ahsan Ali Khoja and
+                  Amy Lee and
+                  Antonio Berlanga-Taylor and
+                  Ashwin Srinath and
+                  bippuspm and
+                  Bret Beheim and
[…]
