+++ README.md
@@ -26,7 +26,13 @@ Each part will be concluded with practical exercises for you to test your skills
 
 Physical copies of this book are not currently available; however, you can download a pdf in the top left corner of this site. Feel free to contribute by reporting a typo or leaving a pull request at <https://github.com/TrevorFrench/R-for-Data-Analysis>.
 
-# Contribution Guide
+## Statement of Need
+
+In the article titled "An empirical study of the rise of big data in business scholarship", the authors suggest that the amount of data that exists in our current society creates a "constant flow of potential new insights for business, government, education and social initiatives" [(Frizzo-Barker et al., 2016)](https://github.com/TrevorFrench/R-for-Data-Analysis/blob/main/paper/paper.bib). This presents an opportunity to educate practitioners in both industry and academia on programmatic data analysis techniques. These practitioners may have historically relied on specialists and/or methodologists to perform analyses, but it is important to ensure that analysis tools are as accessible as the data has become.
+
+There are plenty of resources aimed at teaching specialists how to apply advanced analytics techniques to their chosen discipline; however, there is a notable lack of resources which aim to educate the general public on programmatic data analysis. This phenomenon was observed in an article titled "What is Statistics?" when the authors proclaimed "statistical education has not been sufficiently accessible." [(Brown et al., 2009)](https://github.com/TrevorFrench/R-for-Data-Analysis/blob/main/paper/paper.bib). Furthermore, the contents of R for Data Analysis [(French, 2022)](https://github.com/TrevorFrench/R-for-Data-Analysis/blob/main/paper/paper.bib) are centered around the idea of the "process of data analysis" broadly applied to any discipline. This differs from other high-quality resources, such as "R for Reproducible Scientific Analysis" [(Zimmerman et al., 2019)](https://github.com/TrevorFrench/R-for-Data-Analysis/blob/main/paper/paper.bib), which teaches similar topics in the context of the scientific process.
+
+## Contribution Guide
 
 - You can fix typos, spelling mistakes, or grammatical errors in the documentation directly using the GitHub web interface. Make sure to include a brief description of the changes you are proposing.
 - For other suggestions or larger problems, you can create an "issue" [here](https://github.com/TrevorFrench/R-for-Data-Analysis/issues).
