### 2023-04-08 TrevorFrench/R-for-Data-Analysis#18 @tomsing1
I was thinking adding short "About me" section of your content would be nice.
