### 2023-04-08 TrevorFrench/R-for-Data-Analysis#19 @TrevorFrench
I agree this wasn't adequately addressed in the paper. I made an adjustment in PR #20. Here is the updated [statement of need](https://github.com/TrevorFrench/R-for-Data-Analysis/blob/main/paper/paper.md). Let me know if you think that is sufficient.
