### 2023-04-08 TrevorFrench/R-for-Data-Analysis#14 @TrevorFrench
Hey @corybrunson- thanks for agreeing to review!

I just updated the README in PR #20 to include the statement of need. Let me know if you think that's alright.
