### 2023-04-10 TrevorFrench/R-for-Data-Analysis#16 @tomsing1
Thank you for adding the guidelines to the README!
