### 2023-04-15 #202 @corybrunson
@stats-tgeorge i'm working through the text and have a positive overall impression of the instructional / pedagogical items on the reviewer checklist. I have one thought on it now, and if it's OK with you i'll add any additional thoughts on it to this comment as i go.

* While it hasn't been an issue in any online texts / educational materials i've used so far, i think it's important that they exhibit consistency in style—including tone, educational method, relationship between lessons and examples, etc. Maybe deviations are important in some cases, but in general having a checklist item for consistency would make sense to me. Or maybe consistency should be part of the last item.
