### 2023-04-15 TrevorFrench/R-for-Data-Analysis#17 @TrevorFrench
@tomsing1 Yes, it was intended for self-study. Should I make that more clear in the "Structure of the Book"? I have several anecdotes of it being used in classroom settings; however, I haven't been able to verify the extent/success of those implementations.
