### 2023-04-16 TrevorFrench/R-for-Data-Analysis#18 @TrevorFrench
@tomsing1 I added an "About Me" section to the README as well as the introduction in PR #21
