### 2023-04-17 TrevorFrench/R-for-Data-Analysis#18 @tomsing1
Awesome! Thank you for adding it - really interesting to read about your journey.
