### 2023-04-17 #202 @tomsing1
@stats-tgeorge My review is complete, and the author has addressed all of the concerns I had raised in [their github repository](https://github.com/TrevorFrench/R-for-Data-Analysis/issues). I was able to confirm that the paper matches the criteria of the checklist, except for (maybe) one:

  * The author confirmed that this resource is mainly intended for self-study [here](https://github.com/TrevorFrench/R-for-Data-Analysis/issues/17)

Can you advise on how to proceed, please, e.g. whether the author should discuss this in the paper, or if bullet point in the check list can be ignored given the intended use of the resource?

I think this is a useful resource, and recommend publication in JOSE. One more thought:

* This learning resource takes on a very large scope - including a basic introduction to R, introductory statistics, plotting, report generation, etc. It currently doesn't tackle a full, real-world examples of an actual Data Analysis, which would add additional value (in my opinion). Such content is not a requirement for publication (according to the checklist / scope of JOSE), but maybe @TrevorFrench could consider it for a future edition.
