### 2023-04-17 #202 @stats-tgeorge
@tomsing1 I would say they should mention that is the intended use somewhere in their resource. I like the authors suggestion for in their "structure of the book" section. Thank you!
