### 2023-04-17 TrevorFrench/R-for-Data-Analysis#22 @corybrunson
possibly misleading phrasing for multiple regression
This example begins [section 16.2 on multiple regression](https://trevorfrench.github.io/R-for-Data-Analysis/p4c2-regression.html#multiple-regression):

> If you had more x variables you wanted to add to your linear model, you could add them just like you would in any other math equation. Here’s an example:
> ```r
> lm(data$y ~ data$x1 + data$x2 - data$x3 * data$x4)
> ```

This reads as misleading to me: R's formulae don't work like arithmetic, as in this case the multiplication sign would be interpreted as the three predictors `x3`, `x4`, and `x3 * x4`, while the subtraction sign would exclude them all from the formula, making the result equivalent to `y ~ x1 + x2`. (I omitted `data$` for clarity.) A new user might expect the model to have three parameters: `x1`, `x2`, and `-x3*x4`. (Or they might expect something else.)

To avoid confusion, i would suggest only explaining and illustrating the use of "+" to build multiple regression models, since that is the only operator used in the formulae that appear later in the section.

_Part of [this JOSE review](https://github.com/openjournals/jose-reviews/issues/202)._
