### 2023-04-17 TrevorFrench/R-for-Data-Analysis#23 @corybrunson
lack of deeper content as a limitation versus a design choice
@tomsing1 brought this up in the review thread and i thought it would be worth asking about directly. I agree that the book is conspicuously lacking in real-world examples, as well as in theoretical background or discussions of good practice, for example, in comparison to other book-length resources. However, this struck me more as a design choice than as a limitation; though it was not explicitly mentioned in the README or the JOSE paper. If it is intentional—for example, if the resource is intended for users with background or other resources for these elements—then i think it would be good to explain this in the documentation, and to link to complementary resources there. (But i'm glad to be rejoined on this.)

_Part of [this JOSE review](https://github.com/openjournals/jose-reviews/issues/202)._
