### 2023-04-17 TrevorFrench/R-for-Data-Analysis#23 @corybrunson
Actually, part of this has already been addressed in response to other issues! But i'm still interested in hearing any thoughts.
