### 2023-04-17 TrevorFrench/R-for-Data-Analysis#17 @tomsing1
Sorry, reopening this issue after getting feedback from @stats-tgeorge [here](https://github.com/openjournals/jose-reviews/issues/202#issuecomment-1510672845). He liked your suggestion to describe the intended use of this resource in the "Structure of the Book" section.
