### 2023-04-17 TrevorFrench/R-for-Data-Analysis#17 @TrevorFrench
I added this in PR #24. Let me know if it needs more elaboration. Thanks!
