### 2023-04-19 TrevorFrench/R-for-Data-Analysis#23 @TrevorFrench
@corybrunson yes, this was a choice! I updated the statement of need to include the following:

"Furthermore, the contents of R for Data Analysis (French, 2022) are centered around the idea of the "process of data analysis" broadly applied to any discipline. This differs from other high-quality resources, such as "R for Reproducible Scientific Analysis" (Zimmerman et al., 2019) which teaches similar topics in the context of the scientific process."
