### 2023-05-03 #202 @TrevorFrench
@stats-tgeorge I created a new version/release for this which captures all of the changes that were made through the review process. It's now version 1.1.0. Do I need to do anything to update the version on this issue?
