### 2023-05-04 #202 @stats-tgeorge
@TrevorFrench I looked around for it and I see your recent updates on the repo. There is one open issue still. Did that get resolved just not closed out? Issue 22.
