### 2023-05-04 TrevorFrench/R-for-Data-Analysis#22 @TrevorFrench
Thanks @corybrunson! This was fixed in PR #26
