### 2023-05-04 #202 @TrevorFrench
@stats-tgeorge I just resolved issue 22 and updated the release [here](https://github.com/TrevorFrench/R-for-Data-Analysis/releases/tag/jose).
