### 2023-05-04 #202 @labarba
Next step will be for the author to make an archive on Zenodo and report the DOI here, and the editor to run `@whedon set <doi> as archive`—the author should make sure to edit the archive metadata so that the title and author list match the JOSE paper. Then finally run `@whedon recommend-accept`.
