### 2023-05-20 #202 @labarba
hi @TrevorFrench : please fix the DOI for the Frizzo-Barker reference. The [comment](https://github.com/openjournals/jose-reviews/issues/202#issuecomment-1535378428) from our bot above explains it: remove the URL prefix.

Also, it doesn't quite make sense to cite in the paper the actual learning resource this paper is about. The learning resource will be automatically included as a link in the margin metadata, and also on the landing page of the paper in the JOSE site. You also have a full link at the end of the Summary, which again is redundant to the article metadata linking to the GitHub location of your resource. Consider removing these.
