### 2023-05-20 #202 @TrevorFrench
Hi @labarba -

I removed the doi prefix in this [PR](https://github.com/TrevorFrench/R-for-Data-Analysis/pull/27) and removed the references to the actual resoucre in this [PR](https://github.com/TrevorFrench/R-for-Data-Analysis/pull/28).

Let me know if that looks alright.
